Rename `channel_request` module to `lsps1`
