+++ src/lsps1/channel_manager.rs
